Refactor header styles
This will refactor the header styles to use the golden ratio (sorta) to base the styles of the base font size. This will make it way easier to adjust them in the future. Also, this will multiply the desktop font sizes by 2/3rds to get the mobile font sizes. Pretty coooooool...